requirements.txt
GitPython
python-dotenv
requests